Fixing the Response in Endpoints, Remove the Redundanc code in Repository.
