+++ ECommerceAIMockUp.API/PresentationServiceRegistration.cs
@@ -1,42 +1,9 @@
-﻿using Microsoft.OpenApi.Models;
-
-namespace ECommerceAIMockUp.API
+﻿namespace ECommerceAIMockUp.API
 {
     public static class PresentationServiceRegistration
     {
         public static IServiceCollection AddPresentationServices(this IServiceCollection services)
         {
-            services.AddSwaggerGen(options =>
-            {
-                options.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
-                {
-                    Description = "JWT Authorization header using the bearer scheme.\r\n\r\n" +
-                                  "Enter 'Bearer' [space] and then your token in the text input below.\r\n\r\n" +
-                                  "Example: \"Bearer 12345abcdef\"",
-                    Name = "Authorization",
-                    In = ParameterLocation.Header,
-                    Scheme = "Bearer",
-                    Type = SecuritySchemeType.ApiKey
-                });
-
-                options.AddSecurityRequirement(new OpenApiSecurityRequirement
-    {
-        {
-            new OpenApiSecurityScheme
-            {
-                Reference = new OpenApiReference
-                {
-                    Type = ReferenceType.SecurityScheme,
-                    Id = "Bearer"
-                },
-                Name = "Bearer",
-                In = ParameterLocation.Header
-            },
-            new List<string>()
-        }
-    });
-            });
-
             return services;
         }
     }
