+++ ECommerceAIMockUp.Application/ApplicationServiceRegistration.cs
@@ -1,6 +1,4 @@
-﻿using ECommerceAIMockUp.Application.Services.Implementations;
-using ECommerceAIMockUp.Application.Services.Interfaces.Authentication;
-using Microsoft.Extensions.Configuration;
+﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using System.Reflection;
 
@@ -11,7 +9,6 @@ namespace ECommerceAIMockUp.Application
         public static IServiceCollection AddApplicationServices(this IServiceCollection services, IConfiguration config)
         {
             services.AddAutoMapper(Assembly.GetExecutingAssembly());
-            services.AddScoped<IAuthService, AuthService>();
             return services;
         }
     }
