+++ ECommerceAIMockUp.Application/DTOs/Auth/AuthResponseDto.cs
@@ -2,9 +2,6 @@
 {
     public class AuthResponseDto
     {
-        public string FirstName { get; set; } = default!;
-        public string LastName { get; set; } = default!;
-        public string Email { get; set; }
         public string Token { get; set; } = default!;
     }
 }
