+++ ECommerceAIMockUp.Application/Wrappers/Response.cs
@@ -11,18 +11,9 @@ namespace ECommerceAIMockUp.Application.Wrappers
             IsSucceeded = isSucceeded;
         }
 
-        public Response(T data, HttpStatusCode statusCode, string message, bool isSucceeded)
-        {
-            Data = data;
-            StatusCode = statusCode;
-            Message = message;
-            IsSucceeded = isSucceeded;
-        }
-
         public T Data { get; set; }
 
         public HttpStatusCode StatusCode { get; set; }
-        public string Message { get; set; }
         public bool IsSucceeded { get; set; }
 
     }
