+++ ECommerceAIMockUp.Domain/Entities/AppUser.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.AspNetCore.Identity;
+
+namespace ECommerceAIMockUp.Domain;
+
+public class AppUser : IdentityUser
+{
+    public string FirstName { get; set; } = default!;
+
+    public string LastName { get; set; } = default!;
+
+    public string? City { get; set; }
+
+    public string? Address { get; set; }
+}
