+++ ECommerceAIMockUp.Infrastructure/Configurations/AppUserConfiguration.cs
@@ -1,0 +1,14 @@
+﻿using ECommerceAIMockUp.Domain;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+namespace ECommerceAIMockUp.Infrastructure.Configurations
+{
+    public class AppUserConfiguration : IEntityTypeConfiguration<AppUser>
+    {
+        public void Configure(EntityTypeBuilder<AppUser> builder)
+        {
+
+        }
+    }
+}
